data_wrangling/requirements.txt
appdirs==1.4.4
astroid==2.4.2
certifi==2020.6.20
colorama==0.4.3
distlib==0.3.1
filelock==3.0.12
importlib-metadata==1.7.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.19.1
pandas==1.1.0
pipenv==2020.6.2
pylint==2.5.3
PyMySQL==0.10.0
python-dateutil==2.8.1
python-Levenshtein==0.12.0
pytz==2020.1
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
virtualenv==20.0.30
virtualenv-clone==0.5.4
wrapt==1.12.1
zipp==3.1.0
